+++ src/components/public/LowerNavbar.css
@@ -10,7 +10,6 @@
   display: flex;
   flex-direction: row;
   background-color: white;
-
   width: 100%;
   height: 7.8125rem;
 }
@@ -48,4 +47,7 @@
   height: 1.25rem;
   font-size: 0.875rem;
   line-height:1.25rem;
+}
+.subsitute-lowernavar{
+  height: 9.375rem;
 }
