+++ src/css(subin)/UserProfile.css
@@ -205,4 +205,9 @@ font-size: 1.25rem;
 line-height: 1.9375rem;
 
 }
+.wrap-userinformation{
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
