+++ src/index.css
@@ -3,8 +3,6 @@
   box-sizing: border-box;
   font-family: "Noto Sans KR", sans-serif;
 }
-
-
 body {
   width: 100vw;
   height: 100vh;
