+++ src/pages/UserProfile.js
@@ -1,29 +1,39 @@
 import React, { useEffect } from "react";
-import { useLocation, useNavigate } from "react-router-dom";
+import { useLocation, } from "react-router-dom";
 import axios from "axios";
 //css
 import "../css(subin)/chat.css";
 import "../css(subin)/UserProfile.css";
 //conponent
-import UserList from "../components/userprofile/UserList";
-import ProfileBg from "../components/public/ProfileBg";
 import LowerNavbar from "../components/public/LowerNavbar";
-import MyIcon from "../element/MyIcon";
-import MyModal from "../components/public/MyModal";
+import UserprofileHeader from "../components/userprofile/UserprofileHeader";
+import UserValidationList from "../components/userprofile/UserValidationList";
+import UserInformation from "../components/userprofile/UserInformation";
+import ValidationUserScore from "../components/userprofile/ValidationUserScore";
+import MyModal from"../components/public/MyModal"
+
 
 const UserProfile = () => {
   const [dataProfile, setDataProfile] = React.useState();
-  const [modalIsopen, setmodalIsopen] = React.useState(false);
-  const Auth = sessionStorage.getItem("token");
-  const navigate = useNavigate();
+  const [modalIsopen, setmodalIsopen] = React.useState(true);
+  const roomId = sessionStorage.getItem("roomId");
   const location = useLocation();
   const showUser = location.state;
   const [userList, setUserList] = React.useState([]);
   const [userChoiceValidation, setUserChoiceValidation] =
     React.useState(showUser);
-  const roomId = sessionStorage.getItem("roomId");
+  
   console.log(location);
   console.log(showUser);
+
+  useEffect(() => {
+    showProfileAX();
+  }, [userChoiceValidation]);
+
+  useEffect(() => {
+    showListParticipants();
+  }, [userChoiceValidation]);
+
 
   const handleClickCancel = () => {
     setmodalIsopen(false);
@@ -53,161 +63,23 @@ const UserProfile = () => {
       });
   };
 
-  useEffect(() => {
-    showProfileAX();
-  }, [userChoiceValidation]);
-  useEffect(() => {
-    showListParticipants();
-  }, [userChoiceValidation]);
 
-  const likePlusScore = async () => {
-    const Auth = sessionStorage.getItem("token");
 
-    await axios
-      .post(
-        "http://3.34.129.164/api/likes",
-        { userId: userChoiceValidation },
-        {
-          headers: {
-            Authorization: Auth,
-          },
-        }
-      )
-      .then((res) => {
-        console.log(res);
-        showProfileAX();
-      })
-      .catch((err) => {
-        console.log(err.response.data);
-        alert(err.response.data);
-      });
-  };
 
-  const dislikePlusScore = async () => {
-    const Auth = sessionStorage.getItem("token");
-
-    await axios
-      .post(
-        "http://3.34.129.164/api/hates",
-        { userId: userChoiceValidation },
-        {
-          headers: {
-            Authorization: Auth,
-          },
-        }
-      )
-      .then((res) => {
-        console.log(res); 
-
-        showProfileAX();
-      
-      })
-      .catch((err) => {
-        console.log(err.response);
-        alert(err.response.data);
-      });
-  };
-  const exitPost = async (event) => {
-    event.stopPropagation();
-    await axios
-      .post(`http://3.34.129.164/api/leave/${roomId}`, null, {
-        headers: {
-          Authorization: Auth,
-        },
-      })
-      .then((res) => {
-        console.log(res);
-        sessionStorage.removeItem("rommId");
-        alert("모임나가기가 완료되었습니다!");
-        navigate("/");
-      })
-      .catch((err) => {
-        console.log(err);
-      });
-  };
   return (
     <div className="useprofile-wrap">
-      <div className="userprofile-header-wrap">
-        <div
-          className="userprofile-header-icon"
-          style={{ marginLeft: "28px" }}
-          onClick={(event) => {
-            event.stopPropagation();
-            navigate(-1);
-          }}
-        >
-          <MyIcon sizePx={28} iconName={"arrow_back_ios"} cursor="pointer" />
-
-          
-        </div>
-        <div onClick={exitPost} className="userprofile-header-btn" ><div>모임 나가기</div></div>
-      </div>
-
+      
+      <UserprofileHeader/>
       <div className="wrap-middle">
-        {!modalIsopen ? (
-          <>
-            <div className="user-wrap-countlike">
-              <div className="arrow_box">
-                <MyIcon sizePx={16} iconName={"favorite"} color="beige" />
-                <span className="userprofile-like-text">
-                  {dataProfile?.score}
-                </span>
-              </div>
-            </div>
-            <div className="userprofile-background-img">
-              <ProfileBg ProfileImg={dataProfile?.userProfile} />
-            </div>
-          </>
-        ):<><div className="user-wrap-countlike"></div> <div className="userprofile-background-img"/></>}
-
-        <div className="user-nick font-bold">{dataProfile?.nickname}</div>
-        <div className="user-wrap-sexAndage">
-          <div className="user-sexAndage">
-            <div className="user-sexAndage-text font-medium">
-              {dataProfile?.gender}
-            </div>
-          </div>
-          <div className="user-sexAndage">
-            <div className="user-sexAndage-text font-medium">
-              {dataProfile?.age}
-            </div>
-          </div>
-        </div>
-        <div className="user-wrap-like">
-          <div className="user-like" onClick={dislikePlusScore}>
-            <div className="user-like-text font-bold">-1 싫어요</div>
-          </div>
-          <div
-            className="user-like"
-            style={{ backgroundColor: "#FFBB31" }}
-            onClick={likePlusScore}
-          >
-            <div
-              className="user-like-text font-bold"
-              style={{ color: "white" }}
-            >
-              +1 좋아요
-            </div>
-          </div>
-        </div>
-        <div className="user-wrap-userchange">
-          <div className="user-userchange-title font-bold">
-            모임원들이 평가를 기다리고 있어요!
-          </div>
-          <div className="user-wrap-userchange-userlist">
-            {userList.map((v, i) => (
-              <UserList
-                data={v}
-                key={i}
-                setUserChoiceValidation={setUserChoiceValidation}
-              />
-            ))}
-          </div>
-        </div>
+      <UserInformation dataProfile={dataProfile} modalIsopen={modalIsopen}/>
+        <ValidationUserScore showProfileAX={showProfileAX} userChoiceValidation={userChoiceValidation}/>
+       <UserValidationList userList={userList} setUserChoiceValidation={setUserChoiceValidation}/>
       </div>
-      {/* <MyModal isOpen={modalIsopen} handleClickCancel={handleClickCancel} /> */}
-
-      <LowerNavbar />
+      <MyModal isOpen={modalIsopen} handleClickCancel={handleClickCancel} />
+      <div style="subsitute-lowernavar"></div>
+      <div className="lowernavbar">
+        {!modalIsopen&&<LowerNavbar />}
+      </div>
     </div>
   );
 };
